feat(focusUtils): A simple focus map
A simple focus map that calculates element positions based on mid-pt
co-ordinates.

It assumes that higher level component based focus exists and it is
meant to be used within a complex component that needs it's focus map
compuited on demand.

One should refrain from using this frequently as it takes a performance
hit.

close